Move all BaseHook usages to version_compat in IMAP
Part of https://github.com/apache/airflow/issues/52676
